+++ macros/src/dynify_tests/remote_trait.rs
@@ -1,16 +1,13 @@
 /* This file is @generated for testing purpose */
-trait Trait {
-    async fn test(&self, arg: &str);
-}
 #[allow(async_fn_in_trait)]
 #[allow(clippy::type_complexity)]
-trait DynTrait {
+trait DynTestRemoteTrait {
     fn test<'this, 'arg, 'dynify>(
         &'this self,
         arg: &'arg str,
     ) -> ::dynify::r#priv::Fn<
         (::dynify::r#priv::RefSelf, &'arg str),
-        dyn 'dynify + ::core::future::Future<Output = ()>,
+        dyn 'dynify + ::core::future::Future<Output = usize>,
     >
     where
         'this: 'dynify,
@@ -18,20 +15,21 @@ trait DynTrait {
         Self: 'dynify;
 }
 #[allow(clippy::type_complexity)]
-impl<TraitImplementor: Trait> DynTrait for TraitImplementor {
+impl<TestRemoteTraitImplementor: dynify::r#priv::TestRemoteTrait> DynTestRemoteTrait
+for TestRemoteTraitImplementor {
     fn test<'this, 'arg, 'dynify>(
         &'this self,
         arg: &'arg str,
     ) -> ::dynify::r#priv::Fn<
         (::dynify::r#priv::RefSelf, &'arg str),
-        dyn 'dynify + ::core::future::Future<Output = ()>,
+        dyn 'dynify + ::core::future::Future<Output = usize>,
     >
     where
         'this: 'dynify,
         'arg: 'dynify,
         Self: 'dynify,
     {
-        ::dynify::__from_fn!([self] TraitImplementor::test, self, arg,)
+        ::dynify::__from_fn!([self] TestRemoteTraitImplementor::test, self, arg,)
     }
 }
 fn main() {}